+++ geocoder/src/androidMain/kotlin/tech/kotlinlang/geocoder/GeocoderHelper.android.kt
@@ -1,0 +1,8 @@
+package tech.kotlinlang.geocoder
+
+import tech.kotlinlang.permission.PermissionInitiation
+
+actual fun getGeocoderHelper(): GeocoderHelper {
+    val permissionInitiation = PermissionInitiation.getInstance()
+    return AndroidGeocoderHelper(permissionInitiation.getActivity())
+}