+++ geocoder/src/commonMain/kotlin/tech/kotlinlang/geocoder/GeocoderData.kt
@@ -1,0 +1,7 @@
+package tech.kotlinlang.geocoder
+
+data class GeocoderData(
+    val city: String?,
+    val state: String?,
+    val country: String?,
+)