+++ permission/src/commonMain/kotlin/tech/kotlinlang/permission/location/LocationHelper.kt
@@ -2,4 +2,5 @@ package tech.kotlinlang.permission.location
 
 interface LocationHelper {
     suspend fun fetchLastKnownLocation(): LocationRequestResult
+    suspend fun fetchCurrentLocation(): LocationRequestResult
 }